solution5 -  Days of the month
This code asks the user for a month number and displays how many days are in that month. It uses a dictionary to store the number of days for each month. If the user chooses February, the program also asks whether it is a leap year and displays the number of days accordingly. The program also checks for valid month input and then prints the correct number of days in a month.