solution6 - Brute Force Attack
This code checks an entered password by the user against a stored correct password. If the password is wrong, the user gets up to four additional attempts to enter the correct one. If they enter the password correctly, the program prints “Access granted.” If they use all attempts without success, the program displays a message saying the maximum number of attempts has been reached and that it has been reported to the authorities