solution8 - simple search
This code stores a list of names and first checks if the name "Sam" is in that list. Then it asks the user to enter a name to search. The program checks whether the user’s input exists in the list and prints a message saying whether the name was found or not.